requirements.txt
# Core dependencies
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Redis and search
redis==5.0.1
redis-py-cluster==2.1.3
redisearch==2.0.0

# Logging and utilities
loguru==0.7.2
python-dotenv==1.0.1
requests==2.32.3

# API clients
urllib3==2.0.7

# FastAPI and GraphQL
fastapi==0.104.1
uvicorn[standard]==0.24.0
strawberry-graphql[fastapi]==0.215.0
python-multipart==0.0.6

# System monitoring
psutil==5.9.6

# Additional utilities
pydantic==2.5.0